Split data/nps into public and editor specific modules
